requirements.txt
# ============================================================
# Academic Research Planner — Dependencies
# ============================================================
# Grouped by concern. Do NOT add a library without a
# corresponding entry in DECISIONS.md explaining why.
# ============================================================

# ------------------------------------------------------------
# Web Framework
# ------------------------------------------------------------
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0
python-multipart>=0.0.9          # multipart file uploads

# ------------------------------------------------------------
# Agent Framework
# Compatibility line pinned to pre-1.0 ecosystem due provider package constraints.
# Keep these ranges mutually compatible with langchain-google-genai and tool imports.
# See DECISIONS.md D-005 and D-012
# ------------------------------------------------------------
langchain>=0.3.25,<0.4.0
langchain-core>=0.3.63,<0.4.0
langchain-community>=0.3.24,<0.4.0
langgraph>=0.2.76,<0.3.0
langgraph-checkpoint-postgres>=3.0.0,<4.0.0   # AsyncShallowPostgresSaver

# ------------------------------------------------------------
# LLM Providers
# Gemini + Groq only (free tier access)
# See DECISIONS.md D-012
# ------------------------------------------------------------
langchain-google-genai>=2.0.0,<3.0.0
langchain-groq>=0.2.5,<0.3.0

# ------------------------------------------------------------
# Embeddings
# Cohere Embed v4 — multilingual + multimodal (text + image)
# ------------------------------------------------------------
langchain-cohere>=0.3.0,<0.4.0
cohere>=5.0.0,<6.0.0

# ------------------------------------------------------------
# Vector Store
# Qdrant — one collection per thread
# ------------------------------------------------------------
qdrant-client>=1.11.0,<2.0.0

# ------------------------------------------------------------
# Web Scraping + Search
# Firecrawl — scrape() at ingestion, search() during chat
# ------------------------------------------------------------
firecrawl-py>=1.0.0,<2.0.0

# ------------------------------------------------------------
# Ingestion — PDF, DOCX, PPTX
# PyMuPDF for PDF text + page extraction
# python-docx for DOCX
# python-pptx for PPTX
# ------------------------------------------------------------
PyMuPDF>=1.24.0,<2.0.0
python-docx>=1.1.0,<2.0.0
python-pptx>=1.0.0,<2.0.0
unstructured>=0.15.0             # fallback for complex docs

# ------------------------------------------------------------
# Language Detection
# langdetect: pure Python, works on 3.12 without C++ compilation
# Note: fasttext-langdetect has build issues on Python 3.12 —
# see DECISIONS.md D-010 for rationale
# ------------------------------------------------------------
langdetect>=1.0.9

# ------------------------------------------------------------
# Infrastructure
# asyncpg: async Postgres driver (required by checkpointer)
# psycopg: sync Postgres driver (LangGraph internals)
# pydantic-settings: typed config from .env
# python-dotenv: .env file loading
# ------------------------------------------------------------
asyncpg>=0.29.0,<0.30.0
psycopg[binary]>=3.1.0,<4.0.0
psycopg-pool>=3.2.0,<4.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# ------------------------------------------------------------
# Observability
# LangSmith: tracing for LangChain/LangGraph + @traceable
# ------------------------------------------------------------
langsmith>=0.1.147,<0.4.0

# ------------------------------------------------------------
# Utilities
# tiktoken: token counting (OpenAI tokenizer, good enough approx)
# httpx: async HTTP client for internal service calls
# pyyaml: loading prompt YAML files
# tenacity: retry logic for external API calls
# ------------------------------------------------------------
tiktoken>=0.7.0,<0.8.0
httpx>=0.27.0,<0.28.0
pyyaml>=6.0.1,<7.0.0
tenacity>=8.5.0,<9.0.0

# ------------------------------------------------------------
# Testing
# ------------------------------------------------------------
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
pytest-mock>=3.14.0,<4.0.0
httpx                            # also used for FastAPI TestClient
